+++ CHANGELOG.md
@@ -9,6 +9,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 - **Breaking**, Managed Database: `ManagedDatabaseUserOpernSearchAccessControl` fields changed to pointers
 - **Breaking**, Managed Database: `ManagedDatabaseUserPGAccessControl` fields changed to pointers
 - **Breaking**, Managed Database: `ManagedDatabaseUserRedisAccessControl` fields changed to pointers
+- **Breaking**, Managed Load Balancer: `LoadBalancerFrontendProperties` field `InboundProxyProtocol` to pointer
+- **Breaking**, Managed Object Storage: `CreateManagedObjectStorageUserAccessKeyRequest` field `Enabled` to pointer
+- **Breaking**, Managed Object Storage: `ModifyManagedObjectStorageUserAccessKeyRequest` field `Enabled` to pointer
 
 ### Removed
 - **Breaking**, Managed Database: connection related methods in favor of session